ProgettoFinale060623/requirements.txt
flask
paho-mqtt
google-cloud-firestore
google
flask-login
numpy
sklearn
matplotlib